Making migrations table SQL more portable.
The `WITH TIME ZONE` option is not available in MySQL and neither
are indexes on `NULL` / derived tables.

Due to the recent addition of the `serial_id` column, the index for
"exactly one `NULL`" can be replaced with a CHECK that says only
`serial_id = 0` can have a `NULL` previous.